pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "court-stt"
version = "1.0.0"
description = "Real-time multi-speaker court transcription using Azure Speech Services"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    "azure-cognitiveservices-speech>=1.34.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "websockets>=12.0",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "httpx>=0.25.0",
    "ruff>=0.4.0",
]

[project.scripts]
court-stt-server   = "court_stt.cli:run_server"
court-stt-speaker  = "court_stt.cli:run_speaker"
court-stt-merge    = "court_stt.cli:run_merge"
court-stt-test-mic = "court_stt.cli:run_test_mic"

[tool.setuptools.packages.find]
include = ["court_stt*"]

[tool.setuptools.package-data]
court_stt = ["static/**/*"]

[tool.ruff]
line-length = 100
target-version = "py310"
